mesh-service/requirements.txt
fastapi
uvicorn
numpy
scipy
mpi4py
gmsh
requests
psycopg2-binary
tetgen